pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telethon-fancifier"
version = "0.1.0"
description = "Телеграм-демон на Telethon для постобработки исходящих сообщений по модульным правилам"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "telethon>=1.37.0",
  "httpx>=0.27.0",
  "platformdirs>=4.3.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
  "types-requests>=2.32.0.20241016",
]

[project.scripts]
telethon-fancifier = "telethon_fancifier.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
packages = ["telethon_fancifier"]
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["telethon", "telethon.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["telethon_fancifier.core.daemon"]
disable_error_code = ["untyped-decorator"]
